Prune code for prover/verifier after universal task upgrade (#137)
* prune prover and verifier

* prune commits in verifier

* + refine ProvingTask trait
+ prover: export more 'gen_proof` entry and align them with `execute` entry

* update tester

* revert provingtask trait

* wip

* move point_eval to types

* fix "util" typo in types

* extend execute_and_check method

* refactor integration test

* clippy and fmt

* refactoring integration tests

* canonize task generation

* clippy and fmt

* resume optional for sgement len

* fix checking in instance

* fix according to reviews

* don't need to upload chunk exe?

* trigger

* Revert "trigger"

This reverts commit af98682cf811d507f85ba2a66af4179625166fd1.

* test

* prune unused code

* refine unittest in verifier

---------

+++ crates/types/circuit/src/lib.rs
@@ -117,7 +117,7 @@ fn verify_proof(commitment: &ProgramCommitment, public_inputs: &[u32]) {
         native_addr += FIELDS_PER_U32;
     }
     for &x in public_inputs {
-        openvm::io::store_u32_to_native(native_addr, x as u32);
+        openvm::io::store_u32_to_native(native_addr, x);
         native_addr += FIELDS_PER_U32;
     }
     #[cfg(all(target_os = "zkvm", target_arch = "riscv32"))]
